src/stripe_handler/requirements.txt
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
stripe==8.6.0
typing_extensions==4.10.0
urllib3==2.2.1
